Add build, check and release GitHub Workflows
Added a build workflow that builds the Arch Linux package and uploads it
as an artifact. This is a re-usable workflow.

Added a check workflow that uses the build workflow.

Added a release workflow that also uses the build workflow, and
additionally creates a release and uploads the built package as an asset
for that release.