Optimize sb-tasks
Use the 'export' subcommand of taskwarrior instead of reports. Also
replace awk and tr with jq for calculating the number of tasks when that
is done.

Also replace echo with printf in most cases, except for the final
output of the script, which requires the -e flag of echo to interpret
escape codes as colors.